Factor out FakeCsprng into test_util.rs
